OmicsProfiles

Documentation for OmicsProfiles.

OmicsProfiles.mmread

OmicsProfiles.mmwrite

OmicsProfiles.readinfo

OmicsProfiles.mmread — Function

mmread(filename, retcoord::Bool=false)

Read the contents of the Matrix Market file filename into a matrix, which will be either sparse or dense, depending on the Matrix Market format indicated by coordinate (coordinate sparse storage), or array (dense array storage).

Arguments

filename::String: The file to read.

If retcoord is true, the rows, column and value vectors are returned, if it is a sparse matrix, along with the header information.

OmicsProfiles.mmwrite — Method

mmwrite(filename, matrix)

Write a sparse matrix to .mtx file format.

Arguments

filename::String: The file to write.

matrix::SparseMatrixCSC: The sparse matrix to write.

OmicsProfiles.readinfo — Method

readinfo(file)

Read header information on the size and structure from file. The actual data matrix is not parsed.

Arguments

file: The filename or io stream.